+++ Assets/Scripts/Constants.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Constants
+{
+    public const int TICKS_PER_SEC = 30;
+    public const int MS_PER_TICK = 1000 / TICKS_PER_SEC;
+}
