+++ Assets/Scripts/Player.cs
@@ -1,0 +1,88 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Player : MonoBehaviour
+{
+    public int id;
+    public string username;
+    public CharacterController controller;
+    public float gravity = -9.81f;
+    public float moveSpeed = 5f;
+    public float jumpSpeed = 5f;
+
+    private bool[] inputs;
+    private float yVelocity = 0;
+
+    private void Start()
+    {
+        gravity *= Time.fixedDeltaTime * Time.fixedDeltaTime;
+        moveSpeed *= Time.fixedDeltaTime;
+        jumpSpeed *= Time.fixedDeltaTime;
+    }
+
+    public void Initialize(int _id, string _username)
+    {
+        id = _id;
+        username = _username;
+
+        inputs = new bool[5];
+    }
+
+    public void FixedUpdate()
+    {
+        Vector2 _inputDirection = Vector2.zero;
+        if (inputs[0])  // W
+        {
+            _inputDirection.y += 1;
+        }
+        if (inputs[1])  // A
+        {
+            _inputDirection.x -= 1;
+        }
+        if (inputs[2])  // S
+        {
+            _inputDirection.y -= 1;
+        }
+        if (inputs[3])  // D
+        {
+            _inputDirection.x += 1;
+        }
+
+        Move(_inputDirection);
+    }
+
+    private void Move(Vector2 _inputDirection)
+    {
+        Vector3 _moveDirection = transform.right * _inputDirection.x + transform.forward * _inputDirection.y;
+        _moveDirection *= moveSpeed;
+
+        if (controller.isGrounded)
+        {
+            yVelocity = 0f;
+            if (inputs[4])
+            {
+                yVelocity = jumpSpeed;
+            }
+        }
+        yVelocity += gravity;
+
+        _moveDirection.y = yVelocity;
+        controller.Move(_moveDirection);
+
+        ServerSend.PlayerPosition(this);
+
+        /*
+         * NOTE:
+         * Client rotation is client-authoritative. This is done to avoid 'snapping' from overwriting a player's rotation if rotation was server-authoritative.
+         * If it is desired for the client's rotation to also be server-authoritative, then more advanced topics like client prediction will need to be explored.
+         */
+        ServerSend.PlayerRotation(this);
+    }
+
+    public void SetInput(bool[] _inputs, Quaternion _rotation)
+    {
+        inputs = _inputs;
+        transform.rotation = _rotation;
+    }
+}
