+++ Assets/Scripts/ServerSend.cs
@@ -1,0 +1,118 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ServerSend
+{
+    /// <summary>Sends a packet via TCP to a client.</summary>
+    private static void SendTCPData(int _toClient, Packet _packet)
+    {
+        _packet.WriteLength();
+        Server.clients[_toClient].tcp.SendData(_packet);
+    }
+
+    /// <summary>Sends a packet via UDP to a client.</summary>
+    private static void SendUDPData(int _toClient, Packet _packet)
+    {
+        _packet.WriteLength();
+        Server.clients[_toClient].udp.SendData(_packet);
+    }
+
+    /// <summary>Sends a packet via TCP to all clients.</summary>
+    private static void SendTCPDataToAll(Packet _packet)
+    {
+        _packet.WriteLength();
+        for (int i = 1; i <= Server.MaxPlayers; i++)
+        {
+            Server.clients[i].tcp.SendData(_packet);
+        }
+    }
+
+    /// <summary>Sends a packet via TCP to all but one client.</summary>
+    private static void SendTCPDataToAll(int _exceptClient, Packet _packet)
+    {
+        _packet.WriteLength();
+        for (int i = 1; i <= Server.MaxPlayers; i++)
+        {
+            if (i != _exceptClient)
+            {
+                Server.clients[i].tcp.SendData(_packet);
+            }
+        }
+    }
+
+    /// <summary>Sends a packet via UDP to all clients.</summary>
+    private static void SendUDPDataToAll(Packet _packet)
+    {
+        _packet.WriteLength();
+        for (int i = 1; i <= Server.MaxPlayers; i++)
+        {
+            Server.clients[i].udp.SendData(_packet);
+        }
+    }
+
+    /// <summary>Sends a packet via UDP to all but one client.</summary>
+    private static void SendUDPDataToAll(int _exceptClient, Packet _packet)
+    {
+        _packet.WriteLength();
+        for (int i = 1; i <= Server.MaxPlayers; i++)
+        {
+            if (i != _exceptClient)
+            {
+                Server.clients[i].udp.SendData(_packet);
+            }
+        }
+    }
+
+    #region Packets
+    /// <summary>Sends a welcome message to the client.</summary>
+    public static void Welcome(int _toClient, string _msg)
+    {
+        using (Packet _packet = new Packet((int)ServerPackets.welcome))
+        {
+            _packet.Write(_msg);
+            _packet.Write(_toClient);
+
+            SendTCPData(_toClient, _packet);
+        }
+    }
+
+    /// <summary>Spawns a player.</summary>
+    public static void SpawnPlayer(int _toClient, Player _player)
+    {
+        using (Packet _packet = new Packet((int)ServerPackets.spawnPlayer))
+        {
+            _packet.Write(_player.id);
+            _packet.Write(_player.username);
+            _packet.Write(_player.transform.position);
+            _packet.Write(_player.transform.rotation);
+
+            SendTCPData(_toClient, _packet);
+        }
+    }
+
+    /// <summary>Reports a player's position and sends packet to other clients.</summary>
+    public static void PlayerPosition(Player _player)
+    {
+        using (Packet _packet = new Packet((int)ServerPackets.playerPosition))
+        {
+            _packet.Write(_player.id);
+            _packet.Write(_player.transform.position);
+
+            SendUDPDataToAll(_packet);
+        }
+    }
+
+    /// <summary>Reports a player's rotation and sends packet to other clients.</summary>
+    public static void PlayerRotation(Player _player)
+    {
+        using (Packet _packet = new Packet((int)ServerPackets.playerRotation))
+        {
+            _packet.Write(_player.id);
+            _packet.Write(_player.transform.rotation);
+
+            SendUDPDataToAll(_player.id, _packet);
+        }
+    }
+    #endregion
+}
