+++ GameServer.sln
@@ -1,0 +1,20 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{F9F77F06-66D7-EF2A-06E6-48CD1F3E7407}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F9F77F06-66D7-EF2A-06E6-48CD1F3E7407}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F9F77F06-66D7-EF2A-06E6-48CD1F3E7407}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F9F77F06-66D7-EF2A-06E6-48CD1F3E7407}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F9F77F06-66D7-EF2A-06E6-48CD1F3E7407}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
