--- GameServer/GameServer.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30907.101
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GameServer", "GameServer\GameServer.csproj", "{2E778DEC-F610-4333-A2E0-3A2F93D1B87F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2E778DEC-F610-4333-A2E0-3A2F93D1B87F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2E778DEC-F610-4333-A2E0-3A2F93D1B87F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2E778DEC-F610-4333-A2E0-3A2F93D1B87F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2E778DEC-F610-4333-A2E0-3A2F93D1B87F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FE5BC265-69BC-4F49-B37A-4803C12B1F78}
-	EndGlobalSection
-EndGlobal
