--- GameServer/GameServer/Player.cs
@@ -1,76 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Numerics;
-
-namespace GameServer
-{
-    class Player
-    {
-        public int id;
-        public string username;
-
-        public Vector3 position;
-        public Quaternion rotation;
-
-        private float moveSpeed = 5f / Constants.TICKS_PER_SEC;
-        private bool[] inputs;
-
-        public Player(int _id, string _username, Vector3 _spawnPosition)
-        {
-            id = _id;
-            username = _username;
-            position = _spawnPosition;
-            rotation = Quaternion.Identity;
-
-            inputs = new bool[4];
-        }
-
-        public void Update()
-        {
-            Vector2 _inputDirection = Vector2.Zero;
-            if (inputs[0])  // W
-            {
-                _inputDirection.Y += 1;
-            }
-            if (inputs[1])  // A
-            {
-                _inputDirection.X += 1;
-            }
-            if (inputs[2])  // S
-            {
-                _inputDirection.Y -= 1;
-            }
-            if (inputs[3])  // D
-            {
-                _inputDirection.X -= 1;
-            }
-
-            Move(_inputDirection);
-        }
-
-        private void Move(Vector2 _inputDirection)
-        {
-            Vector3 _forward = Vector3.Transform(new Vector3(0, 0, 1), rotation);
-            Vector3 _right = Vector3.Normalize(Vector3.Cross(_forward, new Vector3(0, 1, 0)));
-
-            Vector3 _moveDirection = _right * _inputDirection.X + _forward * _inputDirection.Y;
-            position += _moveDirection * moveSpeed;
-
-            ServerSend.PlayerPosition(this);
-
-            /*
-             * NOTE:
-             * Client rotation is client-authoritative instead of server-authoritative. This is done to avoid snapping when overwriting a player's rotation.
-             * If it is desired for the client's rotation to also be server-authoritative, then more advanced topics like client prediction will need to be explored.
-             */
-            ServerSend.PlayerRotation(this);
-        }
-
-        public void SetInput(bool[] _inputs, Quaternion _rotation)
-        {
-            inputs = _inputs;
-            rotation = _rotation;
-        }
-    }
-}
