--- GameServer/GameServer/Program.cs
@@ -1,42 +1,0 @@
-﻿using System;
-using System.Threading;
-
-namespace GameServer
-{
-    class Program
-    {
-        private static bool isRunning = false;
-
-        static void Main(string[] args)
-        {
-            Console.Title = "Game Server";
-            isRunning = true;
-
-            Thread mainThread = new Thread(new ThreadStart(MainThread));
-            mainThread.Start();
-
-            Server.Start(50, 20521);
-        }
-
-        private static void MainThread()
-        {
-            Console.WriteLine($"Main thread started. Running at {Constants.TICKS_PER_SEC} ticks per second.");
-            DateTime _nextLoop = DateTime.Now;
-
-            while (isRunning)
-            {
-                while (_nextLoop < DateTime.Now)
-                {
-                    GameLogic.Update();
-
-                    _nextLoop = _nextLoop.AddMilliseconds(Constants.MS_PER_TICK);
-
-                    if (_nextLoop > DateTime.Now)
-                    {
-                        Thread.Sleep(_nextLoop - DateTime.Now);
-                    }
-                }
-            }
-        }
-    }
-}
